fix: change in the structure by eliminating the resolvers and incorporating the functions in the mutations and queries files
